Add support for kubectx
The `kubectx` tool available on Windows does not provide an interactive mode. It means users have to type a context name instead of choose it from a list. This commit adds autocompletion script based on the `kubectl` script. The script was simplified a lot, because `kubectx` supports only one parameter - the context name.