Add support for kubens
The `kubens` tool available on Windows does not provide an interactive mode. It means users have to type a context name instead of choose it from a list. `kubens` works similar as `kubectx` - returns available values when called without parameters. The `kubectx` completion script is converted to a template which is reused for the `kubens` completion.